// src/lib/kline/types.ts
export interface SymbolInfo {
  ticker: string
  name?: string
  shortName?: string
  exchange?: string
  market?: string
  pricePrecision?: number
  volumePrecision?: number
  priceCurrency?: string
  type?: string
  logo?: string
}

export type Timespan = 'second' | 'minute' | 'hour' | 'day' | 'week' | 'month'

export interface Period {
  timeframe: string;
  multiplier: number;
  timespan: Timespan;
  secs: number;
}

export type PaneInds = {
  name: string,
  inds: string[]
}

export type KData = {
  data: KLineData[],
  lays?: unknown[]
}

export type DatafeedWatchCallback = (param: unknown) => void

export type GetKlineArgs = {
  symbol: SymbolInfo,
  period: Period,
  from: number,
  to: number,
  strategy?: string | null
}

export interface Datafeed {
  getSymbols (): Promise<SymbolInfo[]>
  getHistoryKLineData (args: GetKlineArgs): Promise<KData>
  subscribe (symbol: SymbolInfo, callback: DatafeedWatchCallback): void
  unsubscribe (symbol: SymbolInfo): void
}
// ...
export type BarArr = [number, number, number, number, number, number]